添加了测试：在delete方法中当id为空时重定向到list页面

+++ src/main/java/com/baldurtech/contact/ContactController.java
@@ -93,7 +93,11 @@ public class ContactController {
     
     @RequestMapping(value = "delete", method = RequestMethod.POST)
     public String delete(@RequestParam("id") String id) {
-        contactService.delete(Long.valueOf(id));
-        return "redirect:list";
+        if(id == null) {
+            return "redirect:list";
+        } else {
+            contactService.delete(Long.valueOf(id));
+            return "redirect:list";
+        }      
     }
 }
